docker/requirements.txt
apache-airflow-providers-cncf-kubernetes
pandas
faker
openpyxl
pyyaml

